Move re-declared types to `basic_types`
