+++ core/lib/eth_client/Cargo.toml
@@ -16,7 +16,6 @@ zksync_eth_signer.workspace = true
 zksync_config.workspace = true
 zksync_contracts.workspace = true
 
-serde.workspace = true
 thiserror.workspace = true
 async-trait.workspace = true
 jsonrpsee = { workspace = true, features = [
